// packages/gasless-wallet/src/contracts/types/Migration.ts
export interface MigrationInterface extends utils.Interface {
  functions: {
    "migrate()": FunctionFragment;
    "migrationSingleton()": FunctionFragment;
    "safeSingleton()": FunctionFragment;
  };

  getFunction(
    nameOrSignatureOrTopic: "migrate" | "migrationSingleton" | "safeSingleton"
  ): FunctionFragment;

  encodeFunctionData(functionFragment: "migrate", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "migrationSingleton",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "safeSingleton",
    values?: undefined
  ): string;

  decodeFunctionResult(functionFragment: "migrate", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "migrationSingleton",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "safeSingleton",
    data: BytesLike
  ): Result;

  events: {
    "ChangedMasterCopy(address)": EventFragment;
  };

  getEvent(nameOrSignatureOrTopic: "ChangedMasterCopy"): EventFragment;
}